requirements.txt
opencv-python
Pillow
PyMuPDF
mediapipe
numpy